requirements.txt
PyQt6==6.6.1
lxml==5.1.0
pycryptodome==3.20.0
notebook==7.1.2